Update RockPaperScissors__Practice
Simple rock, paper and scissors game for practice. You can simply copy the code and run it - it takes your input and matches it against a computers to determine who the winner is. Afterwards it will ask you if you want to play again. A generic error handling is put in place.